--- mana/packages/mana-common/.fatherrc.ts
@@ -1,9 +1,0 @@
-export default {
-  platform: 'browser',
-  esm: {
-    output: 'es',
-  },
-  cjs: {
-    output: 'lib',
-  },
-};
