--- mana/packages/mana-common/CHANGELOG.md
@@ -1,247 +1,0 @@
-# @difizen/mana-common
-
-## 0.1.32
-
-### Patch Changes
-
-- [#88](https://github.com/difizen/mana/pull/88) [`efec026`](https://github.com/difizen/mana/commit/efec02698841de45108da24897040b528d2c3ec2) Thanks [@sunshinesmilelk](https://github.com/sunshinesmilelk)! - feat: add onWillStart hook for ApplicationComponent
-
-## 0.1.31
-
-### Patch Changes
-
-- [#79](https://github.com/difizen/mana/pull/79) [`0365365`](https://github.com/difizen/mana/commit/03653659a7400430493903e33f44dc2f048576df) Thanks [@lulusir](https://github.com/lulusir)! - fix: localStorage is not defined.
-
-## 0.1.30
-
-### Patch Changes
-
-- [#76](https://github.com/difizen/mana/pull/76) [`14ea61a`](https://github.com/difizen/mana/commit/14ea61ac7ecc717b3091687b9ea4ac0aaf1e6ecf) Thanks [@BroKun](https://github.com/BroKun)! - Use a consistent version.
-
-- [#78](https://github.com/difizen/mana/pull/78) [`92f3efc`](https://github.com/difizen/mana/commit/92f3efcc9852ff7230d9e029e25fdacd9d529d55) Thanks [@sunshinesmilelk](https://github.com/sunshinesmilelk)! - fix text color for some components
-
-## 0.1.28
-
-### Patch Changes
-
-- [#71](https://github.com/difizen/mana/pull/71) [`24b70a0`](https://github.com/difizen/mana/commit/24b70a036be7f4d85548db720ea5859b5a3e5b41) Thanks [@BroKun](https://github.com/BroKun)! - [syringe] Developing multi-instance data structures becomes easier with auto factory.
-
-## 0.1.27
-
-### Patch Changes
-
-- [#62](https://github.com/difizen/mana/pull/62) [`812915f`](https://github.com/difizen/mana/commit/812915f4681c44982cefa5800d6f11ec27a81a20) Thanks [@sunshinesmilelk](https://github.com/sunshinesmilelk)! - fix(umi-plugin): update mana runtime plugin dir name
-
-## 0.1.26
-
-### Patch Changes
-
-- [#60](https://github.com/difizen/mana/pull/60) [`a23732f`](https://github.com/difizen/mana/commit/a23732f365a76f719ac59f2c87de3920f70b1dd7) Thanks [@zhanba](https://github.com/zhanba)! - fix: support query parse in uri
-
-## 0.1.25
-
-### Patch Changes
-
-- [#58](https://github.com/difizen/mana/pull/58) [`df976cb`](https://github.com/difizen/mana/commit/df976cba6bcf49b979dc001ffe68b4e0953522ff) Thanks [@BroKun](https://github.com/BroKun)! - Theme support empty prefix
-
-## 0.1.24
-
-### Patch Changes
-
-- [#56](https://github.com/difizen/mana/pull/56) [`02f92a6`](https://github.com/difizen/mana/commit/02f92a6da23fc653e52c0192c3547ed956a44a22) Thanks [@BroKun](https://github.com/BroKun)! - Support custom css variable prefix
-
-## 0.1.23
-
-### Patch Changes
-
-- [#53](https://github.com/difizen/mana/pull/53) [`c76098a`](https://github.com/difizen/mana/commit/c76098abd7eeae0ea7885d00402a47e15197e2a0) Thanks [@BroKun](https://github.com/BroKun)! - Fix slot view initialization in some cases.
-
-- [#53](https://github.com/difizen/mana/pull/53) [`c24e597`](https://github.com/difizen/mana/commit/c24e5974abd064d3af1d27f321745daa48183a74) Thanks [@BroKun](https://github.com/BroKun)! - Export color protocol.
-
-- [#53](https://github.com/difizen/mana/pull/53) [`8174a13`](https://github.com/difizen/mana/commit/8174a1355b9aca974d7ba68fc6087aef923b757c) Thanks [@BroKun](https://github.com/BroKun)! - Always serialize when storing.
-
-- [#53](https://github.com/difizen/mana/pull/53) [`06bdadd`](https://github.com/difizen/mana/commit/06bdadd00e143151c0b25f63354dab3579d5ee1f) Thanks [@BroKun](https://github.com/BroKun)! - Use modal service api without full modal item
-
-## 0.1.22
-
-### Patch Changes
-
-- ec0b517: Fix core module compatibility
-
-## 0.1.21
-
-### Patch Changes
-
-- 007af7f: Dark mode support for menu.
-- 32133c3: Add notification feature.
-
-## 0.1.20
-
-### Patch Changes
-
-- 690a2aa: Added umi plug-in, now you can use mana in umi project more easily.
-
-## 0.1.19
-
-### Patch Changes
-
-- 43f310a: Remove global style css
-
-## 0.1.18
-
-### Patch Changes
-
-- 02ef50f: basic styles no longer affect global styles.
-
-## 0.1.17
-
-### Patch Changes
-
-- 1acf757: feat: add onInitialized hooks for ApplicationComponent
-
-## 0.1.16
-
-### Patch Changes
-
-- 84f7cd8: No longer import font-awesome styles through less file.
-
-## 0.1.15
-
-### Patch Changes
-
-- 95eb30c: chore: add father config
-
-## 0.1.14
-
-### Patch Changes
-
-- 1b880fe: fix: attributes in ContributionOptionConfig
-
-## 0.1.13
-
-### Patch Changes
-
-- 76745bd: feat: add ContributionOptionConfig
-
-## 0.1.12
-
-### Patch Changes
-
-- 8eee617: chore: update the peerDependencies about react
-
-## 0.1.11
-
-### Patch Changes
-
-- 6fb1d96: mana-observable: Allows defining properties that should not be observed.
-- 67e2969: Add mana-ui package
-- c1cf9c7: Compatible with disposable-collection old API
-
-## 0.1.10
-
-### Patch Changes
-
-- 9f85eb3: App: No longer use proxy to operate tree nodes to improve performance.
-- 5c2f9f3: Workaround on Illegal invocation errors
-- 1e70032: Fix: Object.prototype methods should not be called directly on the object
-- 4cb40cf: Avoid adjacent dividing lines in menus.
-- f090577: No longer proxies access to forzen objects
-- 9fed876: Allow custom modal rendering logic.
-- 395394b: Observable: Allows users to add custom exclusion rules for observable transformations
-- 36bd4ee: Readonly & non-configurable property should return the actual value
-
-## 0.1.9
-
-### Patch Changes
-
-- b5ec5a9: Fix: Divider display between menu groups.
-- 4d9bc01: 1. Once a menu is activated, clicking it again will no longer deactivate it. 2. Even with proxy objects, the original notifier can be found.
-- 4722513: The file tree can be accessed contextually via the file tree menu
-
-## 0.1.8
-
-### Patch Changes
-
-- 9c87e44: 1. Once a menu is activated, clicking it again will no longer deactivate it. 2. Even with proxy objects, the original notifier can be found.
-
-## 0.1.7
-
-### Patch Changes
-
-- 1650890: add simple mode for uri which not escape & unescape
-
-## 0.1.6
-
-### Patch Changes
-
-- b6bac8e: add vscode uri and add parameters in parse function
-
-## 0.1.5
-
-### Patch Changes
-
-- 530df5c: Get notifier before observable transform.
-
-## 0.1.4
-
-### Patch Changes
-
-- d36870f: 1. Add slot related events. 2. Add implementation of basic stateful views.
-
-## 0.1.3
-
-### Patch Changes
-
-- Performance optimization of view rendering and issue fixing of file tree
-
-## 0.1.2
-
-### Patch Changes
-
-- 1. [observable] Do not track dom element object.
-
-## 0.1.1
-
-### Patch Changes
-
-- 1. [observable] Optimize the handling of frozen objects and some built-in types.
-  2. [observable] Support pausing property change events.
-  3. Support tab props & support multi-parameter method to handle menu status.
-
-## 0.1.0
-
-### Minor Changes
-
-- - [observable] remove `Reactable` and use `notifier` as message trigger.
-  - [observable] better event handling.
-  - [syringe] the syringe module now supports setting dependencies.
-
-### Patch Changes
-
-- c137209: 1. (core) add onActiveChange event in DefaultSlotView. 2. (observable) update
-  Notifier api & increased coverage.
-- c9b0853: 1. The syringe module now supports setting dependencies. 2. Observable has
-  been refactored with fewer entities and better event handling. 3. Fixed some other
-  issues and improved code quality.
-- 3f50036: Initial Version
-
-## 0.0.2-alpha.3
-
-### Patch Changes
-
-- 1. (core) add onActiveChange event in DefaultSlotView.
-  2. (observable) update Notifier api & increased coverage.
-
-## 0.0.2-alpha.2
-
-### Patch Changes
-
-- 1. The syringe module now supports setting dependencies.
-  2. Observable has been refactored with fewer entities and better event handling.
-  3. Fixed some other issues and improved code quality.
-
-## 0.0.2-alpha.1
-
-### Patch Changes
-
-- Initial Version
