--- mana/packages/mana-common/README.md
@@ -1,1 +1,0 @@
-# mana-common
