--- mana/packages/mana-common/babel.config.json
@@ -1,3 +1,0 @@
-{
-  "presets": ["@babel/preset-env", "@babel/preset-typescript"]
-}
