--- mana/packages/mana-common/jest.config.js
@@ -1,3 +1,0 @@
-const configs = require('../../jest.config');
-
-module.exports = configs;
