--- mana/packages/mana-common/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "es",
-    "declarationDir": "es"
-  },
-  "types": ["jest"],
-  "exclude": ["node_modules"],
-  "include": ["src"]
-}
