+++ packages/libro-common/src/mana-common/charCode.ts
