+++ packages/libro-common/src/mana-common/disposable.ts
