+++ packages/libro-common/src/mana-common/iterator.ts
