+++ packages/libro-common/src/mana-common/promise-util.ts
