+++ packages/libro-common/src/mana-common/vscode-uri/index.ts
@@ -3,9 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { URI } from './uri';
-import type { UriComponents } from './uri';
-import { Utils } from './utils';
+import { URI } from './uri.js';
+import type { UriComponents } from './uri.js';
+import { Utils } from './utils.js';
 
 export { URI as VscodeURI, Utils as VscodeUriUtils };
 export type { UriComponents };
